Lots of new files, to flesh out what likely will become commonstoolkit
A rush job to get the bulk of my commons code onto GitHub for use on a new job, as USB thumb drives aren't considered secure in that environment. These were abstracted and slightly cleaned up in Apple TextEdit and have not yet been compiled in Eclipse or IntelliJ or rectified against the Maven POM file to make sure all external dependencies are pulled in. Caveat Emptor. Hopefully I'll have a chance to finalize this code soon, pulling it up to javadocs standards and the like.