Cleaned up and added some forgotten file, i/o, and xml utilities
These were meant to be committed a while back, but they got overlooked as I hadn't yet generalized a couple of the methods to pass in some parameters that were previously hard-wired inside the code logic. The converter for XML to HTML is especially useful. It borrows from more than one source and then goes further with some ideas I came up with on my own. As far as I know, no one else has posted an equivalent facility yet, in a highly reusable form such as this one, but bear in mind that it is specific to the context of having an XSLT present to perform the transformation. It's possible that Apache FOP has a similar facility, but likely it requires a lot more heavy lifting than this one as that library is pretty complex overall.