Added smoothing support and switched to StrictMath
As smoothing is generally applicable, I have cleaned up and published my offering in that category, while also switching to StrictMath fro  Math for most transcendental functions in Java. It's not likely to have a performance cost based on research that I did into how these get implemented in modern versions of Java, but I may switch to Apache Commons Math Version 4 for a lot of things in this library once time allows.