Added the bit packing method and fixed the unpacking
The bit packing method is now available, and the bit unpacking method has been cleaned up, corrected, and made safer.

This is not the final version though, as we need to use bytes vs. chars or alter details of the algorithm accordingly.

The problem is that the C version used unsigned chars, so we need to work in 8-bit vs. 7-bit or 16-bit groupings. The limiting factor is the mask values, so we need to see the best approach for solving this incompatibility in data size.