Update BitPackUtilities.java
Massive cleanup and documentation of the bitUnpack() method, now that all of the original intent from the borrowed C code is clear. Lots of interim variables added, to make it clearer what the size is at each point in the shifting and masking logic. Lots of tricks were required to compensate for Java not having unsigned longs.