Update FileUtilities.java
Removed the wrapper for Java NIO Files.copy() as the one in Apache Commons IO is much better and safer. In its place, added stream-based file copy and file channel approaches, as those are only available as private methods inside Core Java API and as Apache Commons only supports either the source or the target/destination (but not both) being a stream or a file channel.