Create ListViewConverter.java
Add an interface that is especially designed to apply to enums so that they can supply a string representation that is preferred for list views in various GUI toolkits without having to resort to using String as the template substitute for such controls as Combo Boxes and the like. This allows us to use data bindings on the data model to the GUI, so that model-only API's can prepare for GUI contexts without adding actual GUI dependencies to them.