Consistency of terminology and Javadocs for network based stuff
This is an attempt to be more consistent about what is named after Servlets and what is named after Servers. The former is used when the implementation is baked-in and alternate technologies are not supported by the class. The new class and variable names also try to make it clear when HTTP is a prerequisite, as this library does not attempt to be a catch-all for all network protocols and all client/server implementation strategies.