Final review of naming conventions for enum-oriented interfaces
Simple is better, but these new names were resisted at first simply because there is a base class in JavaFX called Labeled, but it is unlikely it would be used in the same calling context as this interface.

Downstream usages so far are in unpublished proprietary application code but I hope to find time to refactor my GitHub libraries very soon to take full advantage of this coding style for enums that have shared fields and methods, which otherwise is impossible to achieve as all enums are final classes.

This post also updates two Apache Commons libraries that have had very important bugfixes and functional growth since last updated in the POM file a couple of years ago.