Finally came up with a good generalized baseline for File Mode enumeration
As enums are final classes, I have long delayed adding a FileMode enum, but after studying the generalities and special-casing in several apps, concluded that most file actions can be modeled in a way that applies across applications without leaving too many holes for each app to fill on its own. The "OTHER" instance allows a downstream application to fork their file handling logic to a proprietary enum or class for the file actions that aren't covered here.