Slight protection against StringTokenizer
StringTokenizer is ancient and deeply flawed, but sometimes we inherit code from others and can't take the time to rewrite from scratch. I was unable to address all of this early Java class's flaws, so I backed out the failed experiments and added code comments for downstream clients to take as advice. I was able to at least slightly improve the code that quotes strings, which is meant to protected against early exit from handling a multiu-word field, but StringTokenizer can't really distinguish between a multi-word quoted string and an empty quoted string so there's no useful logic to add at the back end that doesn't do more harm than good. Better to use Apache Commons CsvParser with the token set to a space.