Update TimeUtilities.java
Had to switch toolkits for measuring elapsed time in a way that makes sense, as Java Date/Time API is calendar-based and can't show more than 24 hours of time in the hours column.

The little-known TimeUnit class from the Java Concurrent API turns out to be the more appropriate utility for elapsed time.